// WebApp/src/app/core/http/dto/element/element.dto.ts
export interface ElementDto {
    id: number;
    position: number;
    name: string;
    weight: number;
    symbol: string;
    density: number;
    meltingPoint: number;
    boilingPoint: number;
    atomicRadius: number;
    isVisible: boolean;
    description: string;
    categoryId: number;
    userId: number;
}